Move debugger utils to its own folder
